+++ packages/core-mobile/app/seedless/services/wallet/SeedlessWallet.ts
@@ -35,7 +35,7 @@ import { isTypedData, isTypedDataV1 } from '@avalabs/evm-module'
 import { stripChainAddress } from 'store/account/utils'
 import { AddressPublicKey, Curve } from 'utils/publicKeys'
 import { findPublicKey } from 'utils/publicKeys'
-import { getAddressDerivationPath } from 'utils/getAddressDerivationPath'
+import { getAddressDerivationPath } from 'services/wallet/utils'
 import CoreSeedlessAPIService from '../CoreSeedlessAPIService'
 import SeedlessService from '../SeedlessService'
 import { SeedlessBtcSigner } from './SeedlessBtcSigner'
