+++ packages/core-mobile/app/services/wallet/utils.ts
@@ -2,6 +2,8 @@ import { Avalanche } from '@avalabs/core-wallets-sdk'
 import { Network } from '@avalabs/core-chains-sdk'
 import { TokenUnit } from '@avalabs/core-utils-sdk'
 import { cChainToken } from 'utils/units/knownTokens'
+import { DerivationPathType, NetworkVMType } from '@avalabs/vm-module-types'
+import ModuleManager from 'vmModule/ModuleManager'
 import {
   AvalancheTransactionRequest,
   BtcTransactionRequest,
@@ -48,4 +50,49 @@ export const addBufferToCChainBaseFee = (
   return adjustedBaseFee.toSubUnit() >= minAvax.toSubUnit()
     ? adjustedBaseFee
     : minAvax
+}
+
+export const getAddressDerivationPath = ({
+  accountIndex,
+  vmType,
+  derivationPathType = 'bip44'
+}: {
+  accountIndex: number
+  vmType: Exclude<NetworkVMType, NetworkVMType.PVM | NetworkVMType.HVM>
+  derivationPathType?: DerivationPathType
+}): string => {
+  let derivationPath: string | undefined
+  switch (vmType) {
+    case NetworkVMType.AVM:
+    case NetworkVMType.CoreEth:
+      derivationPath = ModuleManager.avalancheModule.buildDerivationPath({
+        accountIndex,
+        derivationPathType
+      })[vmType]
+      break
+    case NetworkVMType.EVM:
+      derivationPath = ModuleManager.evmModule.buildDerivationPath({
+        accountIndex,
+        derivationPathType
+      })[vmType]
+      break
+    case NetworkVMType.BITCOIN:
+      derivationPath = ModuleManager.bitcoinModule.buildDerivationPath({
+        accountIndex,
+        derivationPathType
+      })[vmType]
+      break
+    case NetworkVMType.SVM:
+      derivationPath = ModuleManager.solanaModule.buildDerivationPath({
+        accountIndex,
+        derivationPathType
+      })[vmType]
+      break
+  }
+
+  if (!derivationPath) {
+    throw new Error(`Unsupported VM type: ${vmType}`)
+  }
+
+  return derivationPath
 }
